requirements.txt
requests
python-telegram-bot
tenacity